refactor: move password hash and check into new module
